+++ services/state/index.tsx
@@ -1,0 +1,33 @@
+import { configureStore } from "@reduxjs/toolkit";
+import { persistStore, persistReducer } from "redux-persist";
+import storage from "redux-persist/lib/storage";
+import { rootReducer, middleware } from "./root";
+import { PersistGate } from "redux-persist/integration/react";
+import { Provider } from "react-redux";
+
+const persistConfig = {
+  key: "root",
+  storage,
+};
+
+const persistedReducer = persistReducer(persistConfig, rootReducer);
+
+export const store = configureStore({
+  reducer: persistedReducer,
+  middleware,
+});
+
+export const persistor = persistStore(store);
+
+type ProviderProps = {
+  children: React.ReactNode;
+};
+export const ReduxProvider: React.FC<ProviderProps> = ({ children }) => {
+  return (
+    <Provider store={store}>
+      <PersistGate loading={null} persistor={persistor}>
+        {children}
+      </PersistGate>
+    </Provider>
+  );
+};
