+++ services/state/ng-eso-api.e2e.test.ts
@@ -1,5 +1,5 @@
 import "@testing-library/jest-dom";
-import { store } from "./index";
+import { store } from "./root";
 import { ngEsoApi } from "./ng-eso-api";
